+++ tests/oihana/signal/MessageTest.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace tests\oihana\signal;
+
+use oihana\signals\Message;
+
+use PHPUnit\Framework\TestCase;
+use stdClass;
+
+class MessageTest extends TestCase
+{
+    public function testCanInstantiateMessage()
+    {
+        $notice = new Message(
+            type: 'info',
+            text: 'hello world' ,
+            target: new stdClass(),
+            context: ['foo' => 'bar']
+        );
+
+        $this->assertSame('info', $notice->type);
+        $this->assertInstanceOf(stdClass::class, $notice->target);
+        $this->assertSame('hello world', $notice->text);
+        $this->assertSame(['foo' => 'bar'], $notice->context);
+    }
+
+    public function testDefaultValues()
+    {
+        $notice = new Message('info' );
+
+        $this->assertSame('info', $notice->type);
+        $this->assertEmpty($notice->text);
+        $this->assertNull($notice->target);
+        $this->assertSame([], $notice->context);
+    }
+}